syncing fork

+++ nucleus/cluster/cli/src/main/java/com/sun/enterprise/admin/cli/cluster/InstallNodeBaseCommand.java
@@ -39,31 +39,25 @@
  */
 package com.sun.enterprise.admin.cli.cluster;
 
-import java.util.logging.Level;
 import com.sun.enterprise.universal.io.SmartFile;
 import com.sun.enterprise.util.SystemPropertyConstants;
 import com.sun.enterprise.util.io.FileListerRelative;
 import com.sun.enterprise.util.io.FileUtils;
 import com.sun.enterprise.util.zip.ZipFileException;
 import com.sun.enterprise.util.zip.ZipWriter;
-import java.io.*;
-import java.net.*;
 import org.glassfish.api.Param;
 import org.glassfish.api.admin.CommandException;
-import javax.inject.Inject;
-
-import org.jvnet.hk2.annotations.Service;
 import org.glassfish.hk2.api.PerLookup;
 import org.glassfish.hk2.api.ServiceLocator;
 import org.glassfish.internal.api.Globals;
+import org.jvnet.hk2.annotations.Service;
 
-import java.io.File;
-import java.io.IOException;
-import java.net.InetAddress;
+import javax.inject.Inject;
+import java.io.*;
 import java.util.ArrayList;
 import java.util.Arrays;
-import java.util.Iterator;
 import java.util.List;
+import java.util.logging.Level;
 
 /**
  * @author Rajiv Mordani
@@ -111,18 +105,14 @@ abstract class InstallNodeBaseCommand extends NativeRemoteCommandsBase {
             precopy();
             zipFile = createZipFileIfNeeded(binDirFiles);
             copyToHosts(zipFile, binDirFiles);
-        }
-        catch (CommandException e) {
+        } catch (CommandException e) {
             throw e;
-        }
-        catch (Exception e) {
+        } catch (Exception e) {
             throw new CommandException(e);
-        }
-        finally {
+        } finally {
             if (!save && delete) {
-                if (zipFile != null) {
-                    if (!zipFile.delete())
-                        zipFile.deleteOnExit();
+                if (zipFile != null && !zipFile.delete()) {
+                    zipFile.deleteOnExit();
                 }
             }
         }
@@ -147,17 +137,17 @@ abstract class InstallNodeBaseCommand extends NativeRemoteCommandsBase {
         String baseRootValue = getSystemProperty(SystemPropertyConstants.PRODUCT_ROOT_PROPERTY);
         File installRoot = new File(baseRootValue);
 
-        File zipFileLocation = null;
-        File glassFishZipFile = null;
+        File zipFileLocation;
+        File glassFishZipFile;
 
         if (archive != null) {
             archive = archive.replace('\\', '/');
-            archiveName = archive.substring(archive.lastIndexOf("/") + 1, archive.length());
-            zipFileLocation = new File(archive.substring(0, archive.lastIndexOf("/")));
+            archiveName = archive.substring(archive.lastIndexOf('/') + 1, archive.length());
+            zipFileLocation = new File(archive.substring(0, archive.lastIndexOf('/')));
             glassFishZipFile = new File(archive);
             if (glassFishZipFile.exists() && !create) {
                 if (logger.isLoggable(Level.FINER))
-                    logger.finer("Found " + archive);
+                    logger.log(Level.FINER, "Found {0}", archive);
                 delete = false;
                 return glassFishZipFile;
             }
@@ -173,35 +163,33 @@ abstract class InstallNodeBaseCommand extends NativeRemoteCommandsBase {
             glassFishZipFile = File.createTempFile("glassfish", ".zip", zipFileLocation);
             String filePath = glassFishZipFile.getCanonicalPath();
             filePath = filePath.replaceAll("\\\\", "/");
-            archiveName = filePath.substring(filePath.lastIndexOf("/") + 1, filePath.length());
+            archiveName = filePath.substring(filePath.lastIndexOf('/') + 1, filePath.length());
         }
 
         FileListerRelative lister = new FileListerRelative(installRoot);
         lister.keepEmptyDirectories();
-        String[] files = lister.getFiles();
 
-        List<String> resultFiles1 = Arrays.asList(files);
-        ArrayList<String> resultFiles = new ArrayList<String>(resultFiles1);
+        List<String> files = Arrays.asList(lister.getFiles());
+        List<String> resultFiles = new ArrayList<>();
 
-        if (logger.isLoggable(Level.FINER))
-            logger.finer("Number of files to be zipped = " +
-                                                            resultFiles.size());
+        if (logger.isLoggable(Level.FINER)) {
+            logger.finer("Number of files to be zipped = " + files.size());
+        }
 
-        Iterator<String> iter = resultFiles.iterator();
-        while (iter.hasNext()) {
-            String fileName = iter.next();
-            String fPath = fileName.substring(fileName.lastIndexOf("/") + 1);
+        for (String fileName: files) {
+            String fPath = fileName.substring(fileName.lastIndexOf('/') + 1);
             if (fPath.equals(glassFishZipFile.getName())) {
-                if (logger.isLoggable(Level.FINER))
+                if (logger.isLoggable(Level.FINER)) {
                     logger.finer("Removing file = " + fileName);
-                iter.remove();
+                }
                 continue;
             }
-            if (fileName.contains("domains") || fileName.contains("nodes")) {
-                iter.remove();
-            }
-            else if (isFileWithinBinDirectory(fileName)) {
-                binDirFiles.add(fileName);
+            if (!fileName.contains("domains") && !fileName.contains("nodes")) {
+
+                resultFiles.add(fileName);
+                if (isFileWithinBinDirectory(fileName)) {
+                    binDirFiles.add(fileName);
+                }
             }
         }
 
@@ -214,7 +202,7 @@ abstract class InstallNodeBaseCommand extends NativeRemoteCommandsBase {
 
         ZipWriter writer = new ZipWriter(FileUtils.safeGetCanonicalPath(glassFishZipFile), installRoot.toString(), filesToZip);
         writer.safeWrite();
-        logger.info("Created installation zip " + FileUtils.safeGetCanonicalPath(glassFishZipFile));
+        logger.log(Level.INFO, "Created installation zip {0}", FileUtils.safeGetCanonicalPath(glassFishZipFile));
 
         return glassFishZipFile;
     }
@@ -240,8 +228,9 @@ abstract class InstallNodeBaseCommand extends NativeRemoteCommandsBase {
 
         char[] buffer = new char[4096];
         int n;
-        while ((n = reader.read(buffer)) >= 0)
+        while ((n = reader.read(buffer)) >= 0) {
             sw.write(buffer, 0, n);
+        }
 
         return sw.toString();
     }
